requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
jose
passlib
httpx
pytest
pytest-asyncio